Added logger, changed CMakeLists

+++ src/asset-baker/asset_main.cpp
@@ -51,7 +51,7 @@ void extract_mesh_from_obj(std::vector<tinyobj::shape_t> &shapes,
     size_t index_offset = 0;
     for (size_t f = 0; f != shapes[s].mesh.num_face_vertices.size(); ++f) {
       // Harcode loading to triangles
-      int fv = 3;
+      size_t fv = 3;
 
       // Loop over vertices in the face
       for (size_t v = 0; v != fv; ++v) {
